open the email in grunt connect task
The connect and live reload tasks were just opening the directory and relying on the file being named "index.html" for it to be served by the watch task. Instead use the values specified in the config so you can use live reload on files with names other than index.html.